refactor: helper to delete all objects in a bucket
