+++ google/cloud/storage/benchmarks/CMakeLists.txt
@@ -16,7 +16,7 @@
 
 if (BUILD_TESTING)
 
-    add_library(storage_benchmarks benchmark_utils.h benchmark_utils.cc)
+    add_library(storage_benchmarks benchmark_utils.h benchmark_utils.cc bounded_queue.h)
     target_link_libraries(
         storage_benchmarks
         PUBLIC storage_client
